+++ Images2WavefrontObj/WavefrontObject/Mesh.h
@@ -6,15 +6,15 @@
 
 struct ObjVertexCoords
 {
-	int X;
-    int Y;
-    int Z;
+	float X;
+    float Y;
+    float Z;
 };
 
 struct ObjTextureCoords
 {
-    int U;
-    int V;
+    float U;
+    float V;
 };
 
 struct SEdgePixels
@@ -32,10 +32,10 @@ public:
     Mesh(std::string p_name);
     virtual ~Mesh();
 
-    void AddVertex(int p_x, int p_y, int p_z);
-    void AddNormal(int p_x, int p_y, int p_z);
-    void AddTexture(int p_u, int p_v);
-    void UpdateVertex(int p_vertexIndex, int p_x, int p_y, int p_z);
+    void AddVertex(float p_x, float p_y, float p_z);
+    void AddNormal(float p_x, float p_y, float p_z);
+    void AddTexture(float p_u, float p_v);
+    void UpdateVertex(int p_vertexIndex, float p_x, float p_y, float p_z);
     void AddFace(Material* p_material);
     void MoveFace(int p_submeshVectorIndex, int p_faceVectorIndex, Material* p_material, bool p_deleteOldFace = true);
     void AddFaceIndices(unsigned short p_vertexIndex, unsigned short p_textureIndex = 0, unsigned short p_normalIndex = 0);
